requirements.txt
apache-airflow==2.5.0
GitPython==3.1.30